requirements.txt
Django==5.2.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.3
python-decouple==3.8 